+++ tests_django/tests/testing/test_views.py
@@ -1,4 +1,6 @@
 """Tests for the social_edu_federation Django integration testing views"""
+import urllib.parse
+
 from django.contrib.auth import get_user as auth_get_user
 from django.core.cache import caches
 from django.urls import reverse
@@ -50,10 +52,33 @@ def test_full_login_process(backend_settings, client, live_server, settings):
     assert response["Location"].startswith(f"{live_server}/saml/idp/sso/?SAMLRequest=")
 
     # 2/ Follow the redirection to the SSO
+    user_login_url = response["Location"]
+    # We need to enforce domain/port for live server as it is hardcoded in test client
+    parsed_url = urllib.parse.urlparse(user_login_url)
+    live_server_name, live_server_port = parsed_url.netloc.split(":")
     response = client.get(
-        f'{reverse("social:begin", args=("saml_fer",))}?idp=local-accepting-idp',
-        follow=True,
+        user_login_url,
+        SERVER_NAME=live_server_name,
+        SERVER_PORT=live_server_port,
     )
+
+    # 3/ Fill form and post
+    form = response.context["user_description_form"]
+    assert (
+        form.initial["acs_url"]
+        == f'http://testserver{reverse("social:complete", args=("saml_fer",))}'
+    )
+    data = form.initial
+    data["surname"] = "Doe"
+    data["given_name"] = "John"
+
+    response = client.post(
+        user_login_url,
+        data=data,
+        SERVER_NAME=live_server_name,
+        SERVER_PORT=live_server_port,
+    )
+    assert "user_description_form" not in response.context
 
     # 3/ Fetch the fake SSO response and POST the data to our endpoint
     assert (