+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+- IDP faker, allow to customize the user attributes.
+
 ## [1.0.0] - 2022-06-24
 
 ### Added